alternate players X and O
