+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"C:\\Users\\Silvia\\Documents\\IT\\IT_Academy\\Front_End\\R7_React_II\\r7_tictactoe\\src\\index.js":"1"},{"size":3488,"mtime":1609696868048,"results":"2","hashOfConfig":"3"},{"filePath":"4","messages":"5","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"bmuqbg","C:\\Users\\Silvia\\Documents\\IT\\IT_Academy\\Front_End\\R7_React_II\\r7_tictactoe\\src\\index.js",[]]
+[{"C:\\Users\\Silvia\\Documents\\IT\\IT_Academy\\Front_End\\R7_React_II\\r7_tictactoe\\src\\index.js":"1","C:\\Users\\Silvia\\Documents\\IT\\IT_Academy\\Front_End\\R7_React_II\\r7_tictactoe\\src\\Board.js":"2","C:\\Users\\Silvia\\Documents\\IT\\IT_Academy\\Front_End\\R7_React_II\\r7_tictactoe\\src\\Square.js":"3"},{"size":659,"mtime":1609763764371,"results":"4","hashOfConfig":"5"},{"size":2919,"mtime":1609769382745,"results":"6","hashOfConfig":"5"},{"size":272,"mtime":1609768202261,"results":"7","hashOfConfig":"5"},{"filePath":"8","messages":"9","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"bmuqbg",{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\Silvia\\Documents\\IT\\IT_Academy\\Front_End\\R7_React_II\\r7_tictactoe\\src\\index.js",[],"C:\\Users\\Silvia\\Documents\\IT\\IT_Academy\\Front_End\\R7_React_II\\r7_tictactoe\\src\\Board.js",[],"C:\\Users\\Silvia\\Documents\\IT\\IT_Academy\\Front_End\\R7_React_II\\r7_tictactoe\\src\\Square.js",[]]
