+++ src/Square.js
@@ -1,0 +1,16 @@
+import React from 'react';
+import './index.css';
+
+function Square(props) {
+
+    return(
+        <button
+            className='square'
+            onClick={props.onClick}
+        >
+            {props.value}
+        </button>
+    );
+}
+
+export default Square;
