+++ src/index.js
@@ -1,79 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
+import Board from './Board';
 import './index.css';
-
-function Square(props) {
-    return(
-        <button
-            className='square'
-            onClick={props.onClick}
-        >
-            {props.value}
-        </button>
-    );
-}
-  
-class Board extends React.Component {
-    constructor(props) {
-        super(props);
-        this.state = {
-            squares: Array(9).fill(null),
-            xIsNext: true, //the value of this boolean will be flipped by handleClick method
-        }
-    }
-
-    handleClick(i) {
-        const squares = this.state.squares.slice(); //creates a copy of the array squares
-        if (calculateWinner(squares) || squares[i]){
-            return;
-        }
-        squares[i] = this.state.xIsNext ? 'X' : 'O'; //sets the square i on 'X' or 'O' depending on the value of xIsNext (I am writing on the copy of the array squares)
-        this.setState({
-            squares: squares,
-            xIsNext: !this.state.xIsNext, 
-        }) //replaces the array squares with the copy in the state and flips the value of xIsNext
-    }
-
-    renderSquare(i) {
-        return (
-            <Square
-                value={this.state.squares[i]}
-                onClick={() => this.handleClick(i)}
-            />
-        );
-    }
-
-    render() {
-        const winner = calculateWinner(this.state.squares);
-        let status;
-        if (winner) {
-            status = 'The winner is: ' + winner;
-        } else {
-            status = 'Next player: ' + (this.state.xIsNext ? 'X' : 'O');
-        }
-
-        return (
-            <div>
-                <div className="status">{status}</div>
-                <div className="board-row">
-                    {this.renderSquare(0)}
-                    {this.renderSquare(1)}
-                    {this.renderSquare(2)}
-                </div>
-                <div className="board-row">
-                    {this.renderSquare(3)}
-                    {this.renderSquare(4)}
-                    {this.renderSquare(5)}
-                </div>
-                <div className="board-row">
-                    {this.renderSquare(6)}
-                    {this.renderSquare(7)}
-                    {this.renderSquare(8)}
-                </div>
-            </div>
-        );
-    }
-}
 
 class Game extends React.Component {
     render() {
@@ -89,26 +17,6 @@ class Game extends React.Component {
             </div>
         );
     }
-}
-
-function calculateWinner(squares) {
-    const lines = [ //possible combinations to win the game
-        [0, 1, 2],
-        [3, 4, 5],
-        [6, 7, 8],
-        [0, 3, 6],
-        [1, 4, 7],
-        [2, 5, 8],
-        [0, 4, 8],
-        [2, 4, 6],
-    ];
-    for (let i = 0; i < lines.length; i++){ //given the array squares, checks if there is a winning combination; until found, returns null
-        const [a, b, c] = lines[i];
-        if (squares[a] && squares[a] === squares[b] && squares[a] === squares[c]) {
-            return squares[a];
-        }
-    }
-    return null;
 }
   
   // ========================================
